w3c-reports/12092014/efl/runned-tests/html/syntax/parsing/template/creating-an-element-for-the-token/template-owner-document-expected.txt

CONSOLE MESSAGE: line 110: ReferenceError: Can't find variable: testInIFrame

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: testInIFrame
